move tree to examples/main
