+++ src/main.tsx
@@ -1,15 +1,1 @@
-import global from '@dojo/framework/shim/global';
-
-if (!global.intern) {
-	import('./examples/src/main');
-} else {
-	const tests = (require as any).context('./', true, /\.spec\.ts(x)?$/);
-	const url = new URL(window.location.href);
-	const params = url.searchParams;
-	const widget = params.get('widget');
-	tests.keys().forEach((id: string) => {
-		if (widget && id.indexOf(widget) !== -1) {
-			tests(id);
-		}
-	});
-}
+import('./examples/src/main');
